some minor layout tweaks, enhanced the filter on the autocomplete to always show the data when nothing is input (so you can filter better). Changed BaseballDataFactory to DataService and moved configs/runs to their own classes.
