+++ ui/.jshintrc
@@ -1,5 +1,6 @@
 {
-  "globalstrict": true,
+  "esnext":true,
+  "globalstrict": false,
   "globals": {
     "angular": false,
     "describe": false,
@@ -8,6 +9,7 @@
     "beforeEach": false,
     "afterEach": false,
     "module": false,
-    "inject": false
+    "inject": false,
+    "require": false
   }
 }
