+++ ui/app/components/main/baseball.js
@@ -1,11 +1,9 @@
-'use strict';
-
-
 import BaseballController from './BaseballController';
-import BaseballDataService from './BaseballFactories';
+import BaseballDataService from './BaseballDataService';
 
 var baseballModule = angular.module('app.main', [])
     .controller('BaseballController', BaseballController)
-    .factory('BaseballDataFactory', BaseballDataService.BaseballDataFactory);
+        .filter('BaseballFilter', BaseballController.BaseballFilter)
+    .service('BaseballDataService', BaseballDataService);
 
 export default baseballModule.name;
