+++ ui/app/index.html
@@ -10,7 +10,7 @@
   <meta name="description" content="">
   <meta name="viewport" content="width=device-width, initial-scale=1">
 </head>
-<body ng-app="app" >
+<body ng-app="app" ng-strict-di>
 
   <div layout="column" layout-fill ui-view></div>
