+++ ui/package.json
@@ -1,10 +1,10 @@
 {
-  "name": "1st-base",
+  "name": "BaseballWorkbench",
   "private": true,
-  "version": "0.0.0",
-  "description": "Bro! Do you even Angular?",
-  "repository": "https://github.com/SMxJrz/1stBase",
-  "license": "MIT",
+  "version": "0.0.1",
+  "description": "Welcome to the baseball workbench",
+  "repository": "https://github.com/bryantrobbins/baseball",
+  "license": "",
   "devDependencies": {
     "babel-core": "^6.5.2",
     "babel-loader": "^6.2.4",
@@ -22,6 +22,8 @@
     "karma-jasmine": "*",
     "karma-junit-reporter": "*",
     "karma-sauce-launcher": "*",
+    "ng-annotate": "^1.2.1",
+    "ng-annotate-webpack-plugin": "^0.1.2",
     "raw-loader": "^0.5.1",
     "style-loader": "^0.13.0",
     "url-loader": "^0.5.7",
@@ -34,16 +36,17 @@
     "angular-animate": "*",
     "angular-aria": "*",
     "angular-material": "^1.0.0",
+    "angular-material-data-table": "*",
     "angular-messages": "*",
     "angular-mocks": "*",
     "angular-ui-router": "^0.2.18",
-    "lodash": "^4.5.1",
-    "angular-material-data-table":"*"
+    "lodash": "^4.5.1"
   },
   "scripts": {
     "postinstall": "",
     "prebuild": "npm install",
     "build": "node_modules/.bin/webpack && cp app/index.html dist",
+    "prestart": "npm install",
     "start": "node_modules/.bin/webpack-dev-server --hot --color --content-base app/",
     "pretest": "npm install",
     "test": "karma start karma.conf.js",
