+++ ui/webpack.config.js
@@ -1,5 +1,8 @@
+"use strict";
+
 var path = require("path");
 var webpack = require("webpack");
+var ngAnnotatePlugin = require("ng-annotate-webpack-plugin");
 
 //var BowerWebpackPlugin = require("bower-webpack-plugin");
 
@@ -8,34 +11,38 @@ var libs = [
     'angular-animate',
     'angular-aria',
     'angular-material',
+    'angular-material-data-table',
     'angular-mocks',
-    'angular-ui-router'
-]
+    'angular-ui-router',
+    'lodash'
+];
 
 var plugins = [
+    new ngAnnotatePlugin({
+        add:true
+    }),
     new webpack.optimize.DedupePlugin(),
     new webpack.optimize.OccurrenceOrderPlugin(),
-    new webpack.optimize.CommonsChunkPlugin('common', 'common.js'),
-]
+    new webpack.optimize.CommonsChunkPlugin('common', 'common.js')
+];
 
 module.exports = {
     context: path.resolve(__dirname, 'app'),
     entry: {
         libs:libs,
-        bundle:['babel-polyfill', './app.js']
+        bundle:['babel-polyfill','./app.js']
     },
     output: {
         path: __dirname + "/dist",
         filename: "[name].js"
     },
+    debug:true,
     module: {
         loaders: [
             { test: /\.js$/, loader: "babel!imports?angular", include: /app|test/},
             { test: /\.css$/, loader: "style!css" },
             { test: /\.tpl\.html$/, loader: "raw" },
-            { test: /\.png$/, loader: 'url-loader?limit=100000&mimetype=image/png'},
-
-
+            { test: /\.png$/, loader: 'url-loader?limit=100000&mimetype=image/png'}
         ]
     },
     plugins: plugins
